Updated deezer-python library to my fork
The current 1.5.0 deezer-python library has an issue where it doesn't return the full content of playlists (due to not following `next` URLs provided, see browniebroke/deezer-python#126 and browniebroke/deezer-python#127 for more information). I forked the repo and fixed this issue, but it hasn't yet been merged and pushed to a new version yet, so I changed the dependency from the pypi version to my (working) forked version.